+++ src/frontend/src/flows/verifiableCredentials/index.ts
@@ -4,7 +4,7 @@ import { withLoader } from "$src/components/loader";
 import { showMessage } from "$src/components/message";
 import { showSpinner } from "$src/components/spinner";
 import { fetchDelegation } from "$src/flows/authorize/fetchDelegation";
-import { validateDerivationOrigin } from "$src/flows/authorize/validateDerivationOrigin";
+import { validateDerivationOrigin } from "$src/utils/validateDerivationOrigin";
 import { getAnchorByPrincipal } from "$src/storage";
 import { resolveIssuerCanisterId } from "$src/utils/canisterIdResolution";
 import { AuthenticatedConnection, Connection } from "$src/utils/iiConnection";
